# docs/api/openapi.json
{
 "components": {
  "schemas": {
...
   "DriftEvent": {
    "description": "Inbound drift notification from Team 1. Member 1 re-checks the drift against the data\nbefore anything is adapted; the event alone never changes a model.",
    "properties": {
     "affected_features": {
      "items": {
       "type": "string"
      },
      "title": "Affected Features",
      "type": "array"
     },
     "data_end_time": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Data End Time"
     },
     "data_start_time": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Data Start Time"
     },
     "dataset_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Dataset Id"
     },
     "detected_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Detected At"
     },
     "drift_detected": {
      "default": true,
      "title": "Drift Detected",
      "type": "boolean"
     },
     "drift_metrics": {
      "additionalProperties": {
       "type": "number"
      },
      "title": "Drift Metrics",
      "type": "object"
     },
     "drift_score": {
      "anyOf": [
       {
        "maximum": 1.0,
        "minimum": 0.0,
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Drift Score"
     },
     "drift_type": {
      "anyOf": [
       {
        "enum": [
         "feature",
         "prediction",
         "concept",
         "data_quality"
        ],
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Drift Type"
     },
     "drifted_data_version": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Drifted Data Version"
     },
     "event_id": {
      "anyOf": [
       {
        "maxLength": 128,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "description": "Caller-supplied id; used for idempotency when present.",
      "title": "Event Id"
     },
     "evidence": {
      "$ref": "#/components/schemas/DriftEvidence"
     },
     "model_id": {
      "maxLength": 200,
      "minLength": 1,
      "title": "Model Id",
      "type": "string"
     },
     "model_type": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Model Type"
     },
     "model_version": {
      "anyOf": [
       {
        "maxLength": 50,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "description": "The model version Team 1 observed drifting. If LIVE has since moved to another version, the event is stale and nothing is adapted.",
      "title": "Model Version"
     },
     "severity": {
      "anyOf": [
       {
        "enum": [
         "LOW",
         "MEDIUM",
         "HIGH",
         "CRITICAL"
        ],
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Severity"
     },
     "source_dataset_version": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Dataset Version"
     },
     "timestamp": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Timestamp"
     }
    },
    "required": [
     "model_id"
    ],
    "title": "DriftEvent",
    "type": "object"
   },
   "DriftEvidence": {
    "additionalProperties": true,
    "properties": {
     "feature": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Feature"
     },
     "p_value": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "P Value"
     },
     "statistic": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Statistic"
     }
    },
    "title": "DriftEvidence",
    "type": "object"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "JobResponse": {
    "properties": {
     "attempt": {
      "default": 0,
      "title": "Attempt",
      "type": "integer"
     },
     "cancel_requested": {
      "default": false,
      "title": "Cancel Requested",
      "type": "boolean"
     },
     "created_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Created At"
     },
     "duplicate": {
      "default": false,
      "title": "Duplicate",
      "type": "boolean"
     },
     "error": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     },
     "job_id": {
      "title": "Job Id",
      "type": "string"
     },
     "model_id": {
      "title": "Model Id",
      "type": "string"
     },
     "quarantined": {
      "default": false,
      "title": "Quarantined",
      "type": "boolean"
     },
     "result": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Result"
     },
     "status": {
      "$ref": "#/components/schemas/JobStatus"
     },
     "strategy": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Strategy"
     },
     "tenant": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Tenant"
     },
     "updated_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Updated At"
     },
     "worker_class": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Worker Class"
     }
    },
    "required": [
     "job_id",
     "model_id",
     "status"
    ],
    "title": "JobResponse",
    "type": "object"
   },
   "JobStatus": {
    "description": "Adaptation job states. Allowed moves between them live in core.state_machine.",
    "enum": [
     "RECEIVED",
     "QUEUED",
     "VALIDATING",
     "DATA_PREPARING",
     "EVALUATING_VERSIONS",
     "REUSE_DECISION",
     "DECISION_PENDING",
     "ADAPTING",
     "VALIDATING_CANDIDATE",
     "REGISTERING",
     "PROMOTING",
     "COMPLETED",
     "FAILED",
     "ROLLED_BACK",
     "TIMED_OUT",
     "CANCELLED",
     "ANALYZING",
     "MODEL_COMPARISON",
     "DECISION_MADE"
    ],
    "title": "JobStatus",
    "type": "string"
   },
   "MappedEventsResponse": {
    "description": "What POST /adaptation/events/from/{mapper} did with a monitoring payload: the events its\nmapper found (none when no record matched) and the job each one produced or already had.",
    "properties": {
     "events": {
      "title": "Events",
      "type": "integer"
     },
     "jobs": {
      "items": {
       "$ref": "#/components/schemas/JobResponse"
      },
      "title": "Jobs",
      "type": "array"
     },
     "mapper": {
      "title": "Mapper",
      "type": "string"
     }
    },
    "required": [
     "mapper",
     "events"
    ],
    "title": "MappedEventsResponse",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "O-RAN Model Adaptation Framework",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
  "/api/v1/adaptation/events": {
   "post": {
    "operationId": "submit_event_api_v1_adaptation_events_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DriftEvent"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/JobResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Submit Event",
    "tags": [
     "adaptation"
    ]
   }
  },
  "/api/v1/adaptation/events/from/{mapper}": {
   "post": {
    "description": "A monitoring system's own payload (an Alertmanager webhook, an Evidently report, ...)\nturned into drift events by the mapper named in DRIFT_MAPPERS, each submitted as if POSTed\nto /adaptation/events. The query parameters set fields the payload does not carry. 201 when\nany event made a new job; 200 when every one was a duplicate or none matched.",
    "operationId": "submit_mapped_events_api_v1_adaptation_events_from__mapper__post",
    "parameters": [
     {
      "in": "path",
      "name": "mapper",
      "required": true,
      "schema": {
       "title": "Mapper",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "model_id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 200,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Model Id"
      }
     },
     {
      "in": "query",
      "name": "model_version",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 50,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Model Version"
      }
     },
     {
      "in": "query",
      "name": "dataset_id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 200,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Dataset Id"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "additionalProperties": true,
        "title": "Payload",
        "type": "object"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MappedEventsResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Submit Mapped Events",
    "tags": [
     "adaptation"
    ]
   }
  },
  "/api/v1/adaptation/jobs": {
   "get": {
    "description": "Jobs, newest first (limit omitted: API_PAGINATION_DEFAULT_LIMIT).",
    "operationId": "get_jobs_api_v1_adaptation_jobs_get",
    "parameters": [
     {
      "in": "query",
      "name": "status",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/JobStatus"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "in": "query",
      "name": "model_id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 200,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Model Id"
      }
     },
     {
      "in": "query",
      "name": "tenant",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 100,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Tenant"
      }
     },
     {
      "in": "query",
      "name": "quarantined",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "title": "Quarantined"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maximum": 1000,
         "minimum": 1,
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Limit"
      }
     },
     {
      "in": "query",
      "name": "offset",
      "required": false,
      "schema": {
       "default": 0,
       "minimum": 0,
       "title": "Offset",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Get Jobs Api V1 Adaptation Jobs Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Jobs",
    "tags": [
     "adaptation"
    ]
   }
  },
  "/api/v1/adaptation/jobs/{job_id}": {
   "get": {
    "description": "A job's current status, result or error, and every state transition it went through.",
    "operationId": "get_job_api_v1_adaptation_jobs__job_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "job_id",
      "required": true,
      "schema": {
       "title": "Job Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Get Job Api V1 Adaptation Jobs  Job Id  Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Job",
    "tags": [
     "adaptation"
    ]
   }
  },
  "/api/v1/adaptation/jobs/{job_id}/cancel": {
   "post": {
    "description": "Cancel a job: 200 when it was queued and is now CANCELLED, 202 when it is running and\nits worker stops it at the next checkpoint (JOB_HEARTBEAT_S), 409 when it already ended or\nis registering or promoting.",
    "operationId": "cancel_job_api_v1_adaptation_jobs__job_id__cancel_post",
    "parameters": [
     {
      "in": "path",
      "name": "job_id",
      "required": true,
      "schema": {
       "title": "Job Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/JobResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Cancel Job",
    "tags": [
     "adaptation"
    ]
   }
  },
...
 }
}